Makefile

+++ lib/maze_storage.cpp
@@ -1,23 +1,19 @@
 #include <iostream>
+
 #include "fields.cpp"
 #include "storage.cpp"
 
 using maze_storage = storage<field_ptr>;
 
-class maze : public maze_storage
-{
+class maze : public maze_storage {
 public:
-  maze(const std::string &maze_str) : maze_storage(maze_str.length())
-  {
+  maze(const std::string &maze_str) : maze_storage(maze_str.length()) {
     field_ptr cell;
-    for (size_t i = 0; i < square; i++)
-    {
-      for (size_t j = 0; j < square; j++)
-      {
+    for (size_t i = 0; i < square; i++) {
+      for (size_t j = 0; j < square; j++) {
         size_t pos = i * square + j;
         field_ptr cell = to_field(maze_str[pos]);
-        if (cell.get()->to_char() == field_type::ENTRANCE)
-        {
+        if (cell.get()->to_char() == field_type::ENTRANCE) {
           entrance_position = point2d(i, j);
         }
         matrix.push_back(cell);
@@ -25,24 +21,20 @@ public:
     }
   }
 
-  void display(point2d player)
-  {
+  void display(point2d player) {
     std::string buff;
-    for (int i = 0; i < square; i++)
-    {
-      for (int j = 0; j < square; j++)
-      {
-          size_t pos = i * square + j;
+    for (size_t i = 0; i < square; i++) {
+      for (size_t j = 0; j < square; j++) {
+        size_t pos = i * square + j;
 
-          if (player.operator==(point2d(i, j)))
-            buff.push_back(field_type::PLAYER);
-          else
-            buff.push_back(matrix[pos].get()->to_char());
+        if (player.operator==(point2d(i, j)))
+          buff.push_back(field_type::PLAYER);
+        else
+          buff.push_back(matrix[pos].get()->to_char());
       }
-          buff.push_back('\n');
-
+      buff.push_back('\n');
     }
-  std::cout << buff << std::endl;
+    std::cout << buff << std::endl;
   }
 
   point2d entrance_position = point2d(0, 0);
